Unit Test Refactor
Move unit tests into their own folder. Prepare scripts for unit + integration tests.
