+++ .vscode/launch.json
@@ -16,11 +16,14 @@
             ]
         },
         {
-            "name": "Launch Tests",
+            "name": "Launch Unit Tests",
             "type": "extensionHost",
             "request": "launch",
             "runtimeExecutable": "${execPath}",
-            "args": ["--extensionDevelopmentPath=${workspaceRoot}", "--extensionTestsPath=${workspaceRoot}/out/test" ],
+            "args": [
+                "--extensionDevelopmentPath=${workspaceRoot}",
+                "--extensionTestsPath=${workspaceRoot}/out/test/unitTests"
+            ],
             "stopOnEntry": false,
             "sourceMaps": true,
             "outFiles": [
