+++ package.json
@@ -26,7 +26,11 @@
     "vscode:prepublish": "tsc -p ./",
     "compile": "tsc -p ./ && gulp tslint",
     "watch": "tsc -watch -p ./",
-    "test": "node ./node_modules/vscode/bin/test",
+    "test": "npm-run-all test:unit test:integration",
+    "test:unit": "cross-env CODE_TESTS_PATH=./out/test/unitTests npm run test:runInVsCode",
+    "test:integration": "npm-run-all test:integration:*",
+    "test:runSuiteInVsCode": "cross-env CODE_TESTS_PATH=./out/test/integrationTests CODE_TESTS_WORKSPACE=./test/integrationTests/testAssets/$OSVC_SUITE npm run test:runInVsCode",
+    "test:runInVsCode": "node ./test/runVsCodeTestsWithAbsolutePaths.js",
     "postinstall": "node ./node_modules/vscode/bin/install"
   },
   "dependencies": {
@@ -46,18 +50,24 @@
   },
   "devDependencies": {
     "@types/chai": "^3.4.34",
+    "@types/chai-arrays": "^1.0.2",
     "@types/fs-extra": "4.0.3",
     "@types/mkdirp": "^0.3.29",
     "@types/mocha": "^2.2.32",
     "@types/node": "^6.0.40",
     "@types/semver": "^5.3.30",
     "@types/tmp": "0.0.32",
+    "async-file": "^2.0.2",
     "chai": "^3.5.0",
+    "chai-arrays": "^2.0.0",
+    "chai-fs": "^1.0.0",
+    "cross-env": "^5.1.1",
     "del": "^2.0.2",
     "gulp": "^3.9.1",
     "gulp-mocha": "^2.1.3",
     "gulp-tslint": "^4.3.0",
     "mocha": "^2.3.3",
+    "npm-run-all": "^4.1.1",
     "plist": "^2.0.1",
     "tslint": "^3.15.1",
     "tslint-microsoft-contrib": "^2.0.12",
