+++ test/unitTests/platform.test.ts
@@ -3,8 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { LinuxDistribution } from '../../src/platform';
 import { should } from 'chai';
-import { LinuxDistribution, PlatformInformation } from '../src/platform';
 
 suite("Platform", () => {
     suiteSetup(() => should());
